move new post modal and image picker file
move to modals folder
