+++ frontend/VolunteerOne/screens/Friends/index.js
@@ -4,7 +4,7 @@ import { StyleSheet, Dimensions, ScrollView } from 'react-native';
 import { Block, theme } from 'galio-framework';
 import { Button } from "../../components";
 import NewAnouncementModal from "../../components/Modals/NewAnnouncementModal"
-import NewPostModal from '../../components/NewPostModal';
+import NewPostModal from '../../components/Modals/NewPostModal';
 
 
 // ================================= View Friends Page ================================= //
